Cleaner Imports +++++++++++++++
Reduced the uglyness of imports somewhat.

The namespace of NoiseTools is now
much cleaner as only the functions actually provided by NoiseTools are
imported in the packages __init__.py => the user will not see all the
imports that NoiseTools uses internally.

Furthermore, NoiseTools now avoids import * statements in its code ...